+++ crates/common/src/fetcher/memory.rs
@@ -3,9 +3,9 @@ use std::collections::HashMap;
 pub type RlpEncodedValue = String;
 
 pub struct MemoryFetcher {
-    pub headers: HashMap<usize, String>,
-    pub accounts: HashMap<usize, HashMap<String, String>>,
-    pub storages: HashMap<usize, HashMap<String, HashMap<String, String>>>,
+    pub headers: HashMap<u64, String>,
+    pub accounts: HashMap<u64, HashMap<String, String>>,
+    pub storages: HashMap<u64, HashMap<String, HashMap<String, String>>>,
 }
 
 impl MemoryFetcher {
@@ -20,9 +20,9 @@ impl MemoryFetcher {
     /// Create a memoizer with pre-filled data
     /// * Note: This is used for testing
     pub fn pre_filled_memoizer(
-        headers: HashMap<usize, RlpEncodedValue>,
-        accounts: HashMap<usize, HashMap<String, RlpEncodedValue>>,
-        storages: HashMap<usize, HashMap<String, HashMap<String, String>>>,
+        headers: HashMap<u64, RlpEncodedValue>,
+        accounts: HashMap<u64, HashMap<String, RlpEncodedValue>>,
+        storages: HashMap<u64, HashMap<String, HashMap<String, String>>>,
     ) -> MemoryFetcher {
         MemoryFetcher {
             headers,
@@ -31,11 +31,11 @@ impl MemoryFetcher {
         }
     }
 
-    pub fn get_rlp_header(&self, block_number: usize) -> Option<RlpEncodedValue> {
+    pub fn get_rlp_header(&self, block_number: u64) -> Option<RlpEncodedValue> {
         self.headers.get(&block_number).cloned()
     }
 
-    pub fn get_rlp_account(&self, block_number: usize, account: String) -> Option<RlpEncodedValue> {
+    pub fn get_rlp_account(&self, block_number: u64, account: String) -> Option<RlpEncodedValue> {
         self.accounts
             .get(&block_number)
             .and_then(|accounts| accounts.get(&account).cloned())
@@ -43,7 +43,7 @@ impl MemoryFetcher {
 
     pub fn get_storage_value(
         &self,
-        block_number: usize,
+        block_number: u64,
         account: String,
         slot: String,
     ) -> Option<String> {
@@ -53,22 +53,16 @@ impl MemoryFetcher {
             .and_then(|slots| slots.get(&slot).cloned())
     }
 
-    pub fn set_header(&mut self, block_number: usize, header: RlpEncodedValue) {
+    pub fn set_header(&mut self, block_number: u64, header: RlpEncodedValue) {
         self.headers.insert(block_number, header);
     }
 
-    pub fn set_account(&mut self, block_number: usize, account: String, value: RlpEncodedValue) {
+    pub fn set_account(&mut self, block_number: u64, account: String, value: RlpEncodedValue) {
         let accounts = self.accounts.entry(block_number).or_default();
         accounts.insert(account, value);
     }
 
-    pub fn set_storage(
-        &mut self,
-        block_number: usize,
-        account: String,
-        slot: String,
-        value: String,
-    ) {
+    pub fn set_storage(&mut self, block_number: u64, account: String, slot: String, value: String) {
         let storages = self.storages.entry(block_number).or_default();
         let slots = storages.entry(account).or_default();
         slots.insert(slot, value);
